+++ src/pages/Applicant/EditableProfile/AdditionalKnowledgeForm/container.tsx
@@ -1,10 +1,10 @@
 import React, { Fragment, FunctionComponent } from "react";
-import { SectionsForm, ITranslations, ISectionsForm } from "../SectionsForm";
+import { SectionsFormSection, ITranslations, ISectionsForm } from "../SectionsFormSection";
 import { useTranslations } from "$hooks";
 
 export const AdditionalKnowledgeFormContainer: FunctionComponent<ISectionsForm> = props => {
   const translations = useTranslations<ITranslations>("additionalKnowledgeForm");
   if (!translations) return <Fragment />;
 
-  return <SectionsForm translations={translations} {...props} />;
+  return <SectionsFormSection translations={translations} {...props} />;
 };
