+++ src/pages/Applicant/EditableProfile/WorkExperienceForm/container.tsx
@@ -1,10 +1,10 @@
 import React, { Fragment, FunctionComponent } from "react";
-import { SectionsForm, ITranslations, ISectionsForm } from "../SectionsForm";
+import { SectionsFormSection, ITranslations, ISectionsForm } from "../SectionsFormSection";
 import { useTranslations } from "$hooks";
 
 export const WorkExperienceFormContainer: FunctionComponent<ISectionsForm> = props => {
   const translations = useTranslations<ITranslations>("workExperienceForm");
   if (!translations) return <Fragment />;
 
-  return <SectionsForm translations={translations} {...props} />;
+  return <SectionsFormSection translations={translations} {...props} />;
 };
